requirements.txt
aiosmtplib==2.0.2
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
comm==0.2.0
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dnspython==2.4.2
docopt==0.6.2
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==2.0.1
Faker==20.0.3
fastapi==0.104.1
fastapi-mail==1.4.1
filelock==3.13.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
identify==2.5.32
idna==3.4
iniconfig==2.0.0
ipykernel==6.27.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
nodeenv==1.8.0
orjson==3.9.10
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
Pygments==2.17.2
pymongo==4.6.0
pytest==7.4.3
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.24.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
